feat(options): allow specifying github options
Allow specifying options to pass to the `Github` object that is created
and used to make requests.

Also updating `github` package to version 7.3.2.

Fixed failing unit tests as well.

BREAKING CHANGE:
  Rather than an `auth` object being passed a `config` object now needs
to be passed, which will contain two keys: `auth` and optionally
`github`. The `auth` key will be passed to `conventionalGithubReleaser`
as before, and the `github` key will be passed to `new Github()`, which
has been moved inside the `conventionalGithubReleaser` export.

```json
{
  "title": "Config Object Schema",
  "type": "object",
  "properties": {
    "auth": {
      "type": "object"
    },
    "github": {
      "type": "object"
    }
  }
}
```